[fix/#217] OnboardingAPI 응답 DTO 구조 분리
