+++ Solply/Solply/Network/DTO/Onboarding/ResponseDTO/OnboardingCompleteResponseDTO.swift
@@ -1,0 +1,24 @@
+//
+//  OnboardingCompleteResponseDTO.swift
+//  Solply
+//
+//  Created by sun on 8/26/25.
+//
+
+import Foundation
+
+struct OnboardingCompleteResponseDTO: ResponseModelType {
+    let selectedTownId: Int
+    let selectedTownName: String
+    let persona: String
+    let nickname: String
+
+    func toEntity() -> OnboardingCompleteInfo {
+        OnboardingCompleteInfo(
+            townId: selectedTownId,
+            townName: selectedTownName,
+            persona: persona,
+            nickname: nickname
+        )
+    }
+}
