+++ Solply/Solply/Network/DTO/Onboarding/ResponseDTO/OnboardingPersonaListResponseDTO.swift
@@ -1,0 +1,21 @@
+//
+//  OnboardingPersonaListResponseDTO.swift
+//  Solply
+//
+//  Created by sun on 8/26/25.
+//
+
+import Foundation
+
+struct OnboardingPersonaListResponseDTO: ResponseModelType {
+    let personaList: [PersonaDTO]
+}
+
+struct PersonaDTO: ResponseModelType, Hashable {
+    let personaType: String
+    let description: String
+
+    func toEntity() -> Persona {
+        Persona(type: personaType, description: description)
+    }
+}
